chore(autoloop): S0 — branch + relocate v1 under src/autoloop/v1/
Prep for v2 (three-agent architecture). No behaviour change.

- src/autoloop.ts        → src/autoloop/v1/runner.ts
- src/autoloop-types.ts  → src/autoloop/v1/types.ts
- AutoloopState gains optional run_mode: 'v1' | 'v2'; v1 stamps 'v1' at init
- resolveConfigPath now walks up to find configs/, robust to depth changes
- Update import sites in session-manager / index / __tests__
- Re-exports preserved (AutoloopRunner + types) — no public API break
- tasks/autoloop-v2.md design doc added

Build + lint + format + 487/487 tests pass.
